#4166 specs updated accordingly
