+++ spec/requests/carto/api/overlays_controller_spec.rb
@@ -10,27 +10,83 @@ describe Carto::Api::OverlaysController do
   it_behaves_like 'overlays controllers' do
   end
 
-  before(:all) do
+  describe '#both endpoints tests' do
 
-    # Spec the routes so that it uses the new controller. Needed for alternative routes testing
-    Rails.application.routes.draw do
+    before(:all) do
+      @user = create_user(
+        username: 'test',
+        email:    'client@example.com',
+        password: 'clientex'
+      )
+      @api_key = @user.api_key
 
-      # new controller
-      scope :module => 'carto/api', :format => :json do
-        get '(/user/:user_domain)(/u/:user_domain)/api/v1/viz/:visualization_id/overlays'     => 'overlays#index',    as: :api_v1_visualizations_overlays_index,  constraints: { visualization_id: /[^\/]+/ }
-        get '(/user/:user_domain)(/u/:user_domain)/api/v1/viz/:visualization_id/overlays/:id' => 'overlays#show',     as: :api_v1_visualizations_overlays_show,   constraints: { visualization_id: /[^\/]+/ }
+      host! 'test.localhost.lan'
+    end
+
+    before(:each) do
+      CartoDB::NamedMapsWrapper::NamedMaps.any_instance.stubs(:get => nil, :create => true, :update => true)
+      delete_user_data @user
+      @table = create_table :user_id => @user.id
+    end
+
+    after(:all) do
+      @user.destroy
+    end
+
+
+    let(:params) { { :api_key => @user.api_key, visualization_id: @table.table_visualization.id } }
+
+
+    it "Lists all overlays" do
+
+      existing_overlay_ids = []
+      get_json api_v1_visualizations_overlays_index_url(params) do |response|
+        response.status.should be_success
+        response.body.count.should == 5   # Newly created overlays have this amount of layers
+        existing_overlay_ids = response.body.map { |overlay| overlay['id'] }
       end
 
-      # old controller
-      scope :module => 'api/json', :format => :json do
+      header_overlay = Overlay::Member.new({
+          type: 'header', 
+          visualization_id: params[:visualization_id],
+          order: 1
+        })
+      header_overlay.store
+
+      text_overlay = Overlay::Member.new({
+          type: 'text',
+          visualization_id: params[:visualization_id],
+          order: 2
+        })
+      text_overlay.store
+
+      new_overlay_ids = [header_overlay.id, text_overlay.id]
+
+      get_json api_v1_visualizations_overlays_index_url(params) do |response|
+        response.status.should be_success
+        current_overlay_ids = response.body.map { |overlay| overlay['id'] }
+        response.body.count.should == new_overlay_ids.count + existing_overlay_ids.count
+        # == checks order, while intersection doesn't 
+        (current_overlay_ids & (existing_overlay_ids + new_overlay_ids) == current_overlay_ids).should eq true
       end
 
     end
 
+    it "Gets the details of an overlay" do
+      header_overlay = Overlay::Member.new({
+          type: 'header', 
+          visualization_id: params[:visualization_id]
+        })
+      header_overlay.store
+
+      get_json api_v1_visualizations_overlays_show_url(params.merge(id: header_overlay.id)) do |response|
+        response.status.should be_success
+        response.body[:id].should == header_overlay.id
+        response.body[:visualization_id].should == params[:visualization_id]
+      end
+    end
+
   end
 
-  after(:all) do
-    Rails.application.reload_routes!
-  end
 
 end
